requirements.txt
fastapi
uvicorn
torch
huggingface_hub
scipy
soundfile
kokoro-onnx
python-multipart
pywebview
requests
numpy
pydantic
ebooklib
xhtml2pdf
beautifulsoup4
lxml

